+++ src/main/java/com/vincit/go/task/slack/executor/TaskSlackMessage.java
@@ -1,5 +1,9 @@
 package com.vincit.go.task.slack.executor;
 
+import com.vincit.go.task.slack.model.MarkdownField;
+
+import java.util.Set;
+
 public class TaskSlackMessage {
 
     private final String title;
@@ -7,13 +11,15 @@ public class TaskSlackMessage {
     private final String iconOrEmoji;
     private final String color;
     private final String displayName;
+    private final Set<MarkdownField> markdownIns;
 
-    public TaskSlackMessage(String displayName, String title, String message, String iconOrEmoji, String color) {
+    public TaskSlackMessage(String displayName, String title, String message, String iconOrEmoji, String color, Set<MarkdownField> markdownIns) {
         this.title = title;
         this.message = message;
         this.iconOrEmoji = iconOrEmoji;
         this.color = color;
         this.displayName = displayName;
+        this.markdownIns = markdownIns;
     }
 
     public String getTitle() {
@@ -34,5 +40,9 @@ public class TaskSlackMessage {
 
     public String getDisplayName() {
         return displayName;
+    }
+
+    public Set<MarkdownField> getMarkdownIns() {
+        return markdownIns;
     }
 }